move SSH operations into transport package
